move router to the separate context
