+++ src/components/product-details/ProductDetails.component.tsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react';
 
 import appStyles from '@/App.module.css';
 import { AppContext } from '@/context/AppContext.context.tsx';
+import { RouterContext } from '@/context/Router.context.tsx';
 import type { Product } from '@/interfaces/Product.interface.tsx';
 
 import { ArrowIconComponent } from '../icon/ArrowIcon.component.tsx';
@@ -11,7 +12,8 @@ import { ProductImageCarouselComponent } from '../product-image-carousel/Product
 import styles from './product-details.module.css';
 
 export const ProductDetailsComponent: React.FC<Product> = (productData) => {
-    const { addToCart, backToPreviousRoute, setRoutePath } = useContext(AppContext);
+    const { backToPreviousRoute, setRoutePath } = useContext(RouterContext);
+    const { addToCart } = useContext(AppContext);
 
     const returnToProducts = () => {
         backToPreviousRoute('/products');
